Removed update_check code from model load.
This was a dictionary used to store information about whether the log
pages had been updated and therefore whether the tables on the UI
needed updating.
Now that the codebase is getting more complicated it seems to make sense
to get rid of this additional logic and just reload the database and 
refresh the UI instead.