Updated LogBuilder for new tmac_tools_lib setup.
tmac_tools_lib has had a major re-write in the way that it deals with
tuflow files. This had a big change on the public api and simplified
a lot of the access.

The LogBuilder module has been updated to accomodate these changes and
the code has been simplified and tidied up.

TODO:
- Lots of comments are probably now wrong. Update these.