Added RUN_OPTIONS and TCF_DIR and fixed results relative path bug.
Updated loaders, database and gui with RUN_OPTIONSand TCF_DIR entries
for the RUN tables.

Added code to the LogBuilder to extract the tcf path and place in the
TCF_DIR entry.

Fixed bug where if an absolute path is used for results it breaks the
functionality used to add tcf name to results relative path.