Updated database verions for bugfix and improved failed write errors.
There was a same version database incompatability issue that was
probably caused by not iterating the version after a change. The
version has now be increased to avoid this in future.

Also improved error handling so that if this happens it is shown to the
user rather than just put in the log.