Updated iefresolver output summary to rich text and moved menu item.
The output summary for unfound files now uses Rich Text format to make
it a bit more readable for the user.

The menu item for the ief resolver has been moved from the Settings >
Tools menuu to a new Tools menu. It didn't seem to make sense to have
it under settings.