Extracted New Entry tab into separate widget.
Create a separate widget for the New Entry tab that is now
loaded along with the other widgets at startup.

Now also uses a TreeView instead of multiple tables to
display the loaded data in the single model entry. This makes
it a lot cleaner for the user and vastly improves the UI.

Still lots to do to get this stable and fully functional.