Fixed update to SHIP issues with IefResolver and loader.
Wasn't accessing the the Ief object with the updated methods.

There was also a problem in the tuflow package of the SHIP
library with inconsistent member variable names. This has
been fixed and the loader works properly again.

Also reactivated the header row of the single model entry
tree, so it is possible to expand the columns.