Update to launch app from main function and fixed load error.
Need to put together some unittests for this, which needs to be able to
run it's own QApplication. Until now the MainGui has erroneously
created a reference to QApplication in the constructor and kept a 
member variable pointing to it. The QApplication is now created in the
main method and and launching and showing of the application and GUI
is maintained in that method.
This has meant some changes have been made to the constructor and and
the custom close events as well.

Fixed a small bug in the model loader. When loading an ief file that
had a reference to a tcf file  that wasn't active it would still load
the tcf and entered it into the database. It now looks to see if the 
use of a 2d scheme is active and ignores the tcf if it isn't.

Added the basic setup of some unit tests. They're not really testing
anything yet.