Bugfix 1D only model load and results reporting.
When a 1D model with a deactivated 2d reference was loaded
it would still load as a 1d/2d model because it only checked
for the file reference. It now checks the scheme activation
status as well.

Results reporting has gone back to folder only instead of
adding the tcf name to the end. This is because there can
be additional sub folders of results and this makes it
clearer.