Bugfix issue in IefResolver autoResolveIef function.
Tests identified an issue in the IefResolver's autofind
logic. When searching for a reference file, if a file with
the same name in a higher level directory was found this would
be chosen.

Fix includes some additional checks to look for all files with
the same name and then try and guess the most appropriate one
by comparing the matching suffix lengths with the original
file path.

All tests now pass.