IefResolver updated outputs and bugfixes.
Improved output display. Now includes a summary of all of
the updated ief file paths.

Resolver will continue to try other files if it fails on
one of them. Before it would stop all together.

Fixed a bug in the secondary search approach that meant the
results files could never be found.