Basics of Run Summary tool now built.
Updated gitignore to not include temporary cache data
created by the tool.

Most of the key functionality is now there.

Needs a lot of work on error handling and catching edge cases.
Should probably put some tests together to check for this.

Need to add a menu item to the log to add to the tool. This
will also need an update of the database to add a LOGS entry
to the RUN table.