Updated database and added getRunStatusInfo.
Database now at version 7. This includes three new columns
for the RUN table: LOG_DIR, RUN_STATUS and MB.

Added the getRunStatusInfo function to Controller. This
uses the TCF_DIR entry in the AllLogs RUN dict to locate
the "_ TUFLOW Simulations.log" file and check the status
of a run. If it is complete it will update the RUN_STATUS
and MB entries.

Tests added for new function in test_controller.py.