Added update run status to tables and add to run summary.
Update run status allows for updating the RUN_STATUS and
MB columns of the RUN table. Can either be done individually
on the right-click menu or for all using the menu.

Add to run summary allows for automatically adding a log
entry to the run summary table by the right click menu.