Fixed issues in RunSummary and added tests.
Tests now included for the loadLogEntry method. These
include a range of different tlf file states.

Some issues were identified when writing the tests that have
been fixed in the method.

Updated test_logbuilder as well to fix the tests after
changing the way that the LOG_DIR value is stored.