Updated LogitSettings to store settings in dict.
Some key variables like logging level are still stored in
the LogitSettings class itself, but most other and
probably all future settings are now stored in a 'main'
dict.

This class is also used to store the other widget dict's
and the state of the global_settings.path_holder dict.