Added parent=self to all dialogs + some bugfix's.
The Gui was not being given as a parent to any of the
dialogs. This meant that it could not be always on top and
would get lost behind other applications if the user changed
the active application after opening a file dialog.
This is now resolved.

There was also an issue with the Load model log menu item
having a broken function call to Controller. Now fixed.

Problem with ief resolver dialog also fixed.