Complete new setup for db backend and log view GUI.
The new setup for software to use the peewee database support is
now all in place.

This includes the full redesign of the user interface for
viewing the logs and the custom query inputs.

Still quite fragile and the tests have not yet been updated for
the new setup.

Ready for beta testing.