Added highlighting and ordering to File summary query table.
If highlight new files checbox is ticked the new model files
will be highlighted green in the table. This uses a new
custom_highlight arg in the addrows method of TableWidgetDb.

Also fixed an error stopping the file summary data from
showing the files in the last run in the loop.