Gui improvements and multi model extractor setup.
GUI's have been adapted to remove any minimum and maximum sizes
in the layout to try and fix the issue with dissapearing widgets
on high resolution screens.

Model extractor has been updated to cater for multiple models in
one go. It still needs a lot of clearing up, error checking and
setting up of saved data and settings.